Test Project Create, it's working but won't update live in FE

+++ app/controllers/api/projects_controller.rb
@@ -1,5 +1,4 @@
 class Api::ProjectsController < ApplicationController
-
   before_action :require_logged_in
 
   def create
@@ -7,7 +6,7 @@ class Api::ProjectsController < ApplicationController
     @project.creator_id = current_user.id
     if @project.save
       current_user.projects << @project
-      render :show #, status: 200
+      render :show, status: 200
     else
       render json: @project.errors.full_messages, status: 422
     end
@@ -19,22 +18,28 @@ class Api::ProjectsController < ApplicationController
       render json: ["Project was not found"], status: 404
       return
     end
+    if @project.creator_id != current_user.id
+      render json: ["Project can only be deleted by its creator"], status: 422
+      return
+    end
     if @project.destroy
-      render :show #, status: 200
+      render :show, status: 200
     else
       render json: ["Project could not be removed"], status: 422
     end
   end
 
+  # ! Not used?
   def index
     @projects = current_user.projects
     render :index
   end
 
+  # ! Not used?
   def show
     @project = Project.includes(:tasks).find_by(id: params[:id])
     if @project
-      render :show #, status: 200
+      render :show, status: 200
     else
       render ["Project was not found"], status: 404
     end
